+++ .gitignore
@@ -36,4 +36,5 @@ out/
 ### VS Code ###
 .vscode/
 
+.DS_Store
 application.yml
