+++ src/main/java/sopt/study/testcode/youngkyung/spring/domain/product/Product.java
@@ -1,10 +1,11 @@
-package sopt.study.testcode.spring.domain.product;
+package sopt.study.testcode.youngkyung.spring.domain.product;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import sopt.study.testcode.spring.domain.BaseEntity;
+import sopt.study.testcode.youngkyung.spring.domain.BaseEntity;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -26,4 +27,13 @@ public class Product extends BaseEntity {
     private String name;
 
     private int price;
+
+    @Builder
+    public Product(String productNumber, ProductType type, ProductSellingStatus sellingStatus, String name, int price) {
+        this.productNumber = productNumber;
+        this.type = type;
+        this.sellingStatus = sellingStatus;
+        this.name = name;
+        this.price = price;
+    }
 }
