+++ src/main/java/sopt/study/testcode/youngkyung/spring/domain/product/ProductRepository.java
@@ -1,4 +1,4 @@
-package sopt.study.testcode.spring.domain.product;
+package sopt.study.testcode.youngkyung.spring.domain.product;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
