+++ src/test/java/sopt/study/testcode/youngkyung/unit/beverage/AmericanoTest.java
@@ -3,7 +3,6 @@ package sopt.study.testcode.youngkyung.unit.beverage;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class AmericanoTest {
     @Test
